PR comments, and refactorings
